refactor: view single manga
Simplified the logic and moved certain logic to where they actually belong. Fetches to public API's and handling of non-sensitive data are in universal load functions. Making requests to the database and handling sensitive data are in the server functions.

Better error handling as well since a user can easily request an invalid manga id in the url.